+++ cmd/invites_list.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"text/tabwriter"
 	"time"
 
+	"github.com/juju/ansiterm"
 	"github.com/urfave/cli"
 
 	"github.com/manifoldco/torus-cli/api"
@@ -14,6 +14,7 @@ import (
 	"github.com/manifoldco/torus-cli/errs"
 	"github.com/manifoldco/torus-cli/hints"
 	"github.com/manifoldco/torus-cli/identity"
+	"github.com/manifoldco/torus-cli/ui"
 )
 
 func invitesList(ctx *cli.Context) error {
@@ -23,13 +24,11 @@ func invitesList(ctx *cli.Context) error {
 	}
 
 	client := api.NewClient(cfg)
+	c := context.Background()
 
-	org, err := client.Orgs.GetByName(context.Background(), ctx.String("org"))
+	org, err := getOrgWithPrompt(client, c, ctx.String("org"))
 	if err != nil {
-		return errs.NewExitError("Could not retrieve org information.")
-	}
-	if org == nil {
-		return errs.NewExitError("Org not found.")
+		return err
 	}
 
 	var states []string
@@ -41,7 +40,7 @@ func invitesList(ctx *cli.Context) error {
 
 	invites, err := client.OrgInvites.List(context.Background(), org.ID, states, "")
 	if err != nil {
-		return errs.NewExitError("Failed to retrieve invites, please try again.")
+		return errs.NewErrorExitError("Failed to retrieve invites, please try again.", err)
 	}
 
 	if len(invites) < 1 {
@@ -68,40 +67,56 @@ func invitesList(ctx *cli.Context) error {
 	// Lookup profiles of those who were invited
 	profiles, err := client.Profiles.ListByID(context.Background(), profileIDs)
 	if err != nil {
-		return errs.NewExitError("Failed to retrieve invites, please try again.")
+		return errs.NewErrorExitError("Failed to retrieve invites, please try again.", err)
 	}
 
+	nameByID := make(map[string]string)
 	usernameByID := make(map[string]string)
 	for _, profile := range profiles {
+		nameByID[profile.ID.String()] = profile.Body.Name
 		usernameByID[profile.ID.String()] = profile.Body.Username
 	}
 
 	fmt.Println("")
 	if ctx.Bool("approved") {
-		fmt.Println("Listing approved invitations for the " + ctx.String("org") + " org")
+		fmt.Println("Listing approved invitations for org " + org.Body.Name)
 	} else {
-		fmt.Println("Listing all pending and accepted invitations for the " + ctx.String("org") + " org")
+		fmt.Println("Listing all pending and accepted invitations for org " + org.Body.Name)
 	}
 	fmt.Println("")
 
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 8, ' ', 0)
-	fmt.Fprintln(w, "EMAIL\tUSERNAME\tSTATE\tINVITED BY\tCREATION DATE")
-	fmt.Fprintln(w, " \t \t \t ")
+	w := ansiterm.NewTabWriter(os.Stdout, 2, 0, 3, ' ', 0)
+	fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%s\t%s\n", ui.Bold("Invited E-Mail"), ui.Bold("Name"), ui.Bold("Username"), ui.Bold("State"), ui.Bold("Invited by"), ui.Bold("Creation Date"))
 	for _, invite := range invites {
-		inviter := usernameByID[invite.Body.InviterID.String()]
+		inviter := nameByID[invite.Body.InviterID.String()]
 		if inviter == "" {
 			continue
 		}
 		identity := invite.Body.Email
-		invitee := "-"
+		inviteeName := "-"
+		inviteeUsername := "-"
 		if invite.Body.InviteeID != nil {
-			invitee = usernameByID[invite.Body.InviteeID.String()]
+			inviteeName = nameByID[invite.Body.InviteeID.String()]
+			inviteeUsername = usernameByID[invite.Body.InviteeID.String()]
 		}
-		fmt.Fprintln(w, identity+"\t"+invitee+"\t"+invite.Body.State+"\t"+inviter+"\t"+invite.Body.Created.Format(time.RFC3339))
+		var state string
+		switch invite.Body.State {
+		case "pending":
+			state = ui.Faint("awaiting acceptance")
+		case "accepted":
+			state = ui.Color(ui.Yellow, "awaiting approval")
+		case "approved":
+			state = ui.Color(ui.Green, "approved")
+		default:
+			state = "-"
+		}
+
+		fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%s\t%s\n", identity, inviteeName, ui.Faint(inviteeUsername), state, inviter, invite.Body.Created.Format(time.RFC3339))
+
 	}
 	w.Flush()
 	fmt.Println("")
 
-	hints.Display(hints.InvitesApprove, hints.TeamMembers)
+	hints.Display(hints.InvitesApprove, hints.OrgMembers)
 	return nil
 }