+++ cmd/services.go
@@ -3,9 +3,7 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"strings"
-	"unicode/utf8"
 
 	"github.com/urfave/cli"
 
@@ -14,6 +12,7 @@ import (
 	"github.com/manifoldco/torus-cli/envelope"
 	"github.com/manifoldco/torus-cli/errs"
 	"github.com/manifoldco/torus-cli/identity"
+	"github.com/manifoldco/torus-cli/ui"
 )
 
 func init() {
@@ -39,12 +38,8 @@ func init() {
 				Name:  "list",
 				Usage: "List services for an organization",
 				Flags: []cli.Flag{
-					orgFlag("org to show services for", true),
+					orgFlag("org to show services for", false),
 					projectFlag("project to show services for", false),
-					cli.BoolFlag{
-						Name:  "all",
-						Usage: "Perform command on all projects",
-					},
 				},
 				Action: chain(
 					ensureDaemon, ensureSession, loadDirPrefs, loadPrefDefaults,
@@ -59,15 +54,6 @@ func init() {
 const serviceListFailed = "Could not list services."
 
 func listServicesCmd(ctx *cli.Context) error {
-	if !ctx.Bool("all") {
-		if len(ctx.String("project")) < 1 {
-			return errs.NewUsageExitError("Missing flags: --project", ctx)
-		}
-	} else {
-		if len(ctx.String("project")) > 0 {
-			return errs.NewUsageExitError("Cannot use --project flag with --all", ctx)
-		}
-	}
 
 	cfg, err := config.LoadConfig()
 	if err != nil {
@@ -77,69 +63,30 @@ func listServicesCmd(ctx *cli.Context) error {
 	client := api.NewClient(cfg)
 	c := context.Background()
 
-	// Look up the target org
-	var org *envelope.Org
-	org, err = client.Orgs.GetByName(c, ctx.String("org"))
+	org, err := getOrgWithPrompt(client, c, ctx.String("org"))
 	if err != nil {
-		return errs.NewErrorExitError(serviceListFailed, err)
-	}
-	if org == nil {
-		return errs.NewExitError("Org not found")
+		return err
 	}
 
-	// Identify which projects to list services for
-	var projectID identity.ID
-	var projects []envelope.Project
-	if ctx.Bool("all") {
-		// Pull all projects for the given orgID
-		projects, err = listProjects(&c, client, org.ID, nil)
-		if err != nil {
-			return errs.NewErrorExitError(serviceListFailed, err)
-		}
-
-	} else {
-		// Retrieve only a single project by name
-		projectName := ctx.String("project")
-		projects, err = listProjects(&c, client, org.ID, &projectName)
-		if err != nil {
-			return errs.NewErrorExitError(serviceListFailed, err)
-		}
-		if len(projects) == 1 {
-			projectID = *projects[0].ID
-		} else {
-			return errs.NewExitError("Project not found")
-		}
+	project, err := getProjectWithPrompt(client, c, org, ctx.String("project"))
+	if err != nil {
+		return err
 	}
 
 	// Retrieve services for targeted org and project
-	services, err := listServices(&c, client, org.ID, &projectID, nil)
+	services, err := listServices(&c, client, org.ID, project.ID, nil)
 	if err != nil {
 		return errs.NewErrorExitError(serviceListFailed, err)
 	}
 
-	// Build map of services to project
-	pMap := make(map[string]envelope.Project)
-	for _, project := range projects {
-		pMap[project.ID.String()] = project
-	}
-	sMap := make(map[string][]envelope.Service)
-	for _, service := range services {
-		ID := service.Body.ProjectID.String()
-		sMap[ID] = append(sMap[ID], service)
-	}
-
-	// Build output of projects/services
+	// Build output of projects/envs
 	fmt.Println("")
-	for projectID, project := range pMap {
-		count := strconv.Itoa(len(sMap[projectID]))
-		title := project.Body.Name + " (" + count + ")"
-		fmt.Println(title)
-		fmt.Println(strings.Repeat("-", utf8.RuneCountInString(title)))
-		for _, service := range sMap[projectID] {
-			fmt.Println(service.Body.Name)
-		}
-		fmt.Println("")
+	fmt.Printf("%s\n", ui.Bold("Services"))
+	for _, s := range services {
+		fmt.Printf("%s\n", s.Body.Name)
 	}
+	
+	fmt.Printf("\nProject /%s/%s has (%d) service%s\n", org.Body.Name, project.Body.Name, len(services), plural(len(services)))
 
 	return nil
 }
@@ -256,4 +203,12 @@ func createServiceCmd(ctx *cli.Context) error {
 
 	fmt.Printf("Service %s created.\n", serviceName)
 	return nil
+}
+
+func plural(n int) (string) {
+	if n == 1 {
+		return ""
+	} else {
+		return "s"
+	}
 }